Moved Translation component to Resource
