+++ src/Sylius/Bundle/ContactBundle/DependencyInjection/Configuration.php
@@ -24,7 +24,7 @@ use Sylius\Component\Contact\Model\TopicInterface;
 use Sylius\Component\Contact\Model\TopicTranslation;
 use Sylius\Component\Contact\Model\TopicTranslationInterface;
 use Sylius\Component\Resource\Factory\Factory;
-use Sylius\Component\Translation\Factory\TranslatableFactory;
+use Sylius\Component\Resource\Factory\TranslatableFactory;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
