+++ src/Sylius/Bundle/CoreBundle/Locale/TranslationLocaleProvider.php
@@ -12,7 +12,7 @@
 namespace Sylius\Bundle\CoreBundle\Locale;
 
 use Sylius\Component\Locale\Context\LocaleContextInterface;
-use Sylius\Component\Translation\Provider\LocaleProviderInterface;
+use Sylius\Component\Resource\Provider\LocaleProviderInterface;
 
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
