+++ src/Sylius/Bundle/ProductBundle/DependencyInjection/Compiler/ServicesPass.php
@@ -12,7 +12,7 @@
 namespace Sylius\Bundle\ProductBundle\DependencyInjection\Compiler;
 
 use Sylius\Component\Resource\Factory\Factory;
-use Sylius\Component\Translation\Factory\TranslatableFactory;
+use Sylius\Component\Resource\Factory\TranslatableFactory;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
