+++ src/Sylius/Bundle/TranslationBundle/Resources/config/doctrine/model/AbstractTranslation.orm.xml
@@ -13,6 +13,6 @@
 
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping">
 
-    <mapped-superclass name="Sylius\Component\Translation\Model\AbstractTranslation" />
+    <mapped-superclass name="Sylius\Component\Resource\Model\AbstractTranslation" />
 
 </doctrine-mapping>
