+++ src/Sylius/Component/Archetype/Model/ArchetypeTranslation.php
@@ -11,7 +11,7 @@
 
 namespace Sylius\Component\Archetype\Model;
 
-use Sylius\Component\Translation\Model\AbstractTranslation;
+use Sylius\Component\Resource\Model\AbstractTranslation;
 
 /**
  * @author Gonzalo Vilaseca <gvilaseca@reiss.co.uk>
