+++ src/Sylius/Component/Resource/Model/TranslationInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Component\Translation\Model;
+namespace Sylius\Component\Resource\Model;
 
 /**
  * @author Gonzalo Vilaseca <gvilaseca@reiss.co.uk>
