+++ src/Sylius/Component/Resource/Provider/LocaleProvider.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Component\Translation\Provider;
+namespace Sylius\Component\Resource\Provider;
 
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
