--- src/Sylius/Component/Translation/.gitignore
@@ -1,5 +1,0 @@
-vendor/
-bin/
-
-composer.phar
-composer.lock
