--- src/Sylius/Component/Translation/.travis.yml
@@ -1,18 +1,0 @@
-language: php
-
-sudo: false
-
-php:
-  - 7.0
-  - 5.6
-  - 5.5
-
-before_install:
-    - echo "memory_limit=2048M" >> ~/.phpenv/versions/$(phpenv version-name)/etc/conf.d/travis.ini
-    - phpenv config-rm xdebug.ini || true
-
-install:
-    - composer install --no-interaction --prefer-source
-
-script:
-    - bin/phpspec run -fpretty --verbose
