--- src/Sylius/Component/Translation/composer.json
@@ -1,45 +1,0 @@
-{
-    "name": "sylius/translation",
-    "type": "library",
-    "description": "Translation component for translatable resources.",
-    "keywords": ["translation", "translatable", "resource", "sylius", "ecommerce"],
-    "homepage": "http://sylius.org",
-    "license": "MIT",
-    "authors": [
-        {
-            "name":     "Paweł Jędrzejewski",
-            "homepage": "http://pjedrzejewski.com"
-        },
-        {
-            "name":     "Sylius project",
-            "homepage": "http://sylius.org"
-        },
-        {
-            "name":     "Community contributions",
-            "homepage": "http://github.com/Sylius/Sylius/contributors"
-        }
-    ],
-    "require": {
-        "php": "^5.5.9|^7.0",
-
-        "sylius/resource": "0.18.*"
-    },
-    "require-dev": {
-        "phpspec/phpspec": "^2.4"
-    },
-    "config": {
-        "bin-dir": "bin"
-    },
-    "autoload": {
-        "psr-4": { "Sylius\\Component\\Translation\\": "" }
-    },
-    "autoload-dev": {
-        "psr-4": { "Sylius\\Component\\Translation\\spec\\": "spec/" }
-    },
-    "minimum-stability": "dev",
-    "extra": {
-        "branch-alias": {
-            "dev-master": "0.18-dev"
-        }
-    }
-}
