--- src/Sylius/Component/Translation/phpspec.yml.dist
@@ -1,5 +1,0 @@
-suites:
-    main:
-        namespace: Sylius\Component\Translation
-        psr4_prefix: Sylius\Component\Translation
-        src_path: .
